# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
...
  internal_features_theme_users_transport_http.GetThemeUserResponse:
    properties:
      addition_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      id:
        example: 1
        type: integer
      percentages:
        example: 0
        type: integer
      theme_id:
        example: 3
        type: integer
      user_id:
        example: 4
        type: integer
      version:
        example: 2
        type: integer
    type: object
  internal_features_theme_users_transport_http.PatchThemeUserRequest:
    type: object
  internal_features_theme_users_transport_http.PatchThemeUserResponse:
    properties:
      addition_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      id:
        example: 1
        type: integer
      percentages:
        example: 0
        type: integer
      theme_id:
        example: 3
        type: integer
      user_id:
        example: 4
        type: integer
      version:
        example: 2
        type: integer
    type: object
...
  internal_features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        example: Ivan Ivanov
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+79998887766"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        example: Максим Максимович
        type: string
      phone_number:
        example: "+79998887766"
        type: string
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.UserDTOResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Learning Progress REST-API scheme
  title: Golang My Project
  version: "1.0"
paths:
...
  /themes_user/{id}:
    delete:
      description: Удаление существующей темы пользователя в системе по ее ID
      parameters:
      - description: ID удаляемой темы пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление темы пользователя
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: ThemeUser not found
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
      summary: Удаление темы пользователя
      tags:
      - themes_user
    get:
      description: Получение конкретной темы пользователя в системе по ее ID
      parameters:
      - description: ID получаемой темы пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Тема пользователя успешно найдена
          schema:
            $ref: '#/definitions/internal_features_theme_users_transport_http.GetThemeUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: ThemeUser not found
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
      summary: Получение темы пользователя
      tags:
      - themes_user
    patch:
      consumes:
      - application/json
      description: |-
        Изменение информации об уже существующем в системе теме пользователя
        Ограничение: `completed` не может быть выставлен как null
      parameters:
      - description: ID изменяемой темы пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: PatchThemeUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_theme_users_transport_http.PatchThemeUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное изменение темы пользователя
          schema:
            $ref: '#/definitions/internal_features_theme_users_transport_http.PatchThemeUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: ThemeUser not found
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
      summary: Изменение темы пользователя
      tags:
      - themes_user
  /users:
    get:
      description: Получение списка пользователей с опциональной пагинацией
      parameters:
      - description: Размер страницы с пользователями
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с пользователями
        in: query
        name: offser
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка пользователей
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
      summary: Список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      description: Удаление существующего пользователя в системе по его ID
      parameters:
      - description: ID удаляемого пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление пользователя
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
      summary: Удаление пользователя
      tags:
      - users
    get:
      description: Получение конкретного пользователя в системе по его ID
      parameters:
      - description: ID получаемого пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно найден
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
      summary: Получение пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: |-
        Изменение информации об уже существующем в системе пользователе
        ### Логика обновления полей (Three-state logic):
        1. **Поле не передано**: `"phone_number"` игнорируется, значение в БД не меняется
        2. **Явно передано значение**: `"phone_number": "+71112223344"` - устанавливает новый номер телефона в БД
        3. **Передан null**: `"phone_number": null` - очищает поле в БД (set to NULL)
        Ограничение: `full_name` не может быть выставлен как null
      parameters:
      - description: ID изменяемого пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное изменение пользователя
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_LayMan011_Golang-My-Study_internal_core_transport_http_response.ErrorResponse'
      summary: Изменение пользователя
      tags:
      - users
swagger: "2.0"